requirements.txt
setuptools==65.5.0
gym==0.21.0
# torch==1.11.0
numpy==1.23.5
tensorboard==1.15.0
matplotlib==3.7.1
mani_skill2
pillow==9.5.0
tqdm==4.65.0
scikit-learn==1.2.0
torchvision==0.15.2
timm==0.6.13
pyglet==2.0.7
sklearn